requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
black==23.1.0
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
crispy-bootstrap5==0.7
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-autocomplete-light==3.9.4
django-crispy-forms==2.0
django-environ==0.10.0
django-js-asset==2.0.0
django-mptt==0.14.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
gunicorn==20.1.0
idna==3.4
kombu==5.2.4
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
redis==4.5.1
requests==2.28.2
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
